Shane/shared demo (#1690)
* demo

* more tweaks
